Move leader_skill_header to common
