+++ padinfo/view/common.py
@@ -2,9 +2,12 @@ from typing import TYPE_CHECKING
 
 from discordmenu.embed.base import Box
 from discordmenu.embed.text import Text, BoldText
+from tsutils import char_to_emoji
+from tsutils.enums import LsMultiplier
 from tsutils.query_settings import QuerySettings
 
-from padinfo.common.emoji_map import get_awakening_emoji
+from padinfo.common.emoji_map import get_awakening_emoji, get_emoji
+from padinfo.core.leader_skills import ls_multiplier_text, ls_single_multiplier_text
 from padinfo.view.components.monster.header import MonsterHeader
 
 if TYPE_CHECKING:
@@ -38,3 +41,14 @@ def get_awoken_skill_description(awoken_skill: "AwokenSkillModel"):
 def invalid_monster_text(query: str, monster: "MonsterModel", append_text: str, link=False):
     base_text = 'Your query `{}` found {}{}.'
     return base_text.format(query, MonsterHeader.short_with_emoji(monster, link=link).to_markdown(), append_text)
+
+
+def leader_skill_header(m: "MonsterModel", lsmultiplier: LsMultiplier, transform_base: "MonsterModel"):
+    return Box(
+        BoldText('Leader Skill'),
+        BoldText(ls_multiplier_text(m.leader_skill) if lsmultiplier == LsMultiplier.lsdouble
+                 else char_to_emoji(1) + ' ' + ls_single_multiplier_text(m.leader_skill)),
+        BoldText('(' + get_emoji(
+            '\N{DOWN-POINTING RED TRIANGLE}') + '7x6)') if m != transform_base and transform_base.leader_skill.is_7x6 else None,
+        delimiter=' '
+    )
